requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
Flask-WTF==1.2.2
flask-mongoengine==1.0.0
mongoengine==0.29.1
pymongo==4.10.1
email_validator==2.2.0
python-dotenv==1.0.1
Werkzeug==3.1.3
WTForms==3.2.1
dnspython==2.7.0
